refactor(netcdf): extract variable mutation into engines/variables.py (STR-1)
Move set_variable / add_variable / remove_variable / rename_variable out
of netcdf.py into Variables(_Engine) (issue #615). NetCDF keeps thin
façades delegating to the engine; signatures, behaviour, and return
types are unchanged. The engine reaches the container's GDAL plumbing
(_writable_root_group / _replace_raster / _invalidate_caches /
_get_or_create_dimension / _add_md_array_to_group) through the
weakref-proxied back-reference self._ds.

Wired as self.varops — deliberately NOT 'variables', which is an
existing read-side property returning the lazy variable dict; the
engine must not shadow it. add_variable lazily imports NetCDF for its
isinstance check to avoid the netcdf.py <-> engines.variables cycle.
